pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "pyarrow>=23.0.1",
    "sqlalchemy>=2.0.49",
    "tqdm>=4.67.3",
]

[dependency-groups]
dev = [
    "jupyter>=1.1.1",
    "pgcli>=4.4.0",
]
